moved StorySaveEvents to new package
