+++ stories/src/main/java/com/wordpress/stories/compose/event/StoryLoadEvents.kt
@@ -1,0 +1,13 @@
+package com.wordpress.stories.compose.event
+
+import com.wordpress.stories.compose.story.StoryIndex
+
+class StoryLoadEvents {
+    data class StoryLoadStart(
+        var storyIndex: StoryIndex
+    )
+
+    data class StoryLoadEnd(
+        var storyIndex: StoryIndex
+    )
+}
